pyproject.toml
# pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/plan_execute"]

[project]
name = "plan-execute"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arize-phoenix>=11.18.0",
    "arize-phoenix-otel>=0.12.1",
    "asyncpg>=0.30.0",
    "dspy-ai>=3.0.1",
    "fastapi>=0.115.0",
    "ipython>=9.4.0",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.28",
    "langfuse>=3.2.4",
    "langgraph>=0.6.3",
    "langgraph-checkpoint-postgres>=2.0.23",
    "litellm>=1.55.0",
    "openai>=1.98.0",
    "openinference-instrumentation-langchain>=0.1.50",
    "psycopg-binary>=3.2.9",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.32.0",
]
